email_poc/requirements.txt
jinja2>=3.1
pypdf>=4.0
python-docx>=1.0
requests>=2.28
